requirements.txt
pyserial>=3.5
PyYAML>=6.0
Flask>=2.3.0
